Fix a bug where any students can join teacher with the same code
Students had the ability to join the teacher with
the same code if they wasn't already joined with the
same teacher, this creates a studentTeacher with a changable
studentId.
Eg. A student with a teacher passes his id to another student
who is not with the teacher, now both students will be considered
one student at the teacher db, and they both have access on the
details of that student!!!